Move ConvosVault into ConvosCore, remove separate package
The Vault is deeply coupled to ConvosCore infrastructure (keychain,
InboxStateMachine, InboxLifecycleManager, push notifications,
SessionManager). A separate package required redundant protocol
bridges (VaultServiceProtocol, VaultMessageProcessorProtocol) just
to avoid the dependency — adding complexity without benefit.

Moved into ConvosCore/Sources/ConvosCore/Vault/:
- ContentTypes/ (DeviceKeyBundle, DeviceKeyShare, DeviceRemoved codecs)
- VaultClient.swift (XMTP wrapper, lifecycle handling)
- VaultManager.swift (key sync orchestration)
- PairingCoordinator.swift (confirmation code gen/validation)
- VaultServiceProtocol.swift (merged with VaultSessionIntegration)
- VaultSessionIntegration.swift (VaultManager conformance)

Consolidated protocols:
- VaultServiceProtocol now includes start/stop/pause/resume/unpair/shareNewKey
- Removed VaultKeyShareNotifier (replaced by VaultServiceProtocol.shareNewKey)
- Removed ConversationKeyInfo (replaced by InboxKeyInfo)
- Removed VaultSessionIntegration protocol (merged into VaultServiceProtocol)

Tests moved to ConvosCoreTests:
- VaultContentTypeTests (25 tests)
- VaultManagerTests (8 tests)
- PairingCoordinatorTests (14 tests)

386 ConvosCore tests passing (339 + 47 from ConvosVault), full app builds.
