+++ ConvosCore/Sources/ConvosCore/Vault/PairingCoordinator.swift
