+++ ConvosCore/Sources/ConvosCore/Vault/VaultServiceProtocol.swift
@@ -1,4 +1,5 @@
 import Foundation
+@preconcurrency import XMTPiOS
 
 public struct InboxKeyInfo: Sendable {
     public let inboxId: String
@@ -15,6 +16,10 @@ public struct InboxKeyInfo: Sendable {
 }
 
 public protocol VaultServiceProtocol: Sendable {
+    func startVault(signingKey: SigningKey, options: ClientOptions) async throws
+    func stopVault()
+    func pauseVault()
+    func resumeVault() async
     func unpairSelf() async throws
     func shareNewKey(_ keyInfo: InboxKeyInfo) async
 }
