+++ ConvosCore/Tests/ConvosCoreTests/VaultContentTypeTests.swift
@@ -1,4 +1,4 @@
-@testable import ConvosVault
+@testable import ConvosCore
 import Foundation
 import Testing
 import XMTPiOS
